initial commit of the updated bridge
