+++ sm_cnn/.gitignore
@@ -1,5 +1,6 @@
 *pyc
-trec_eval-8.0/trec_eval
-data/
+*.pt
+text/
 trained_models/
 trec_eval-8.0/trec_eval.dSYM
+data/
