+++ sm_cnn/create_dataset.sh
