+++ sm_cnn/eval/gdeval.pl
