+++ sm_cnn/evaluate.py
