+++ sm_cnn/model.py
@@ -1,92 +1,79 @@
 import torch
 import torch.nn as nn
+
 import torch.nn.functional as F
 
-# logging setup
-import logging
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
+class SmPlusPlus(nn.Module):
+    def __init__(self, config):
+        super(SmPlusPlus, self).__init__()
+        output_channel = config.output_channel
+        questions_num = config.questions_num
+        answers_num = config.answers_num
+        words_dim = config.words_dim
+        filter_width = config.filter_width
+        self.mode = config.mode
 
-ch = logging.StreamHandler()
-ch.setLevel(logging.DEBUG)
-formatter = logging.Formatter('%(levelname)s - %(message)s')
-ch.setFormatter(formatter)
-logger.addHandler(ch)
+        n_classes = config.target_class
+        ext_feats_size = 4
 
-
-class QAModel(nn.Module):
-
-    @staticmethod
-    def save(model, model_fname):
-        torch.save(model, model_fname)
-
-
-    @staticmethod
-    def load(model_fname):
-        return torch.load(model_fname)
-
-
-    def __init__(self, input_n_dim, filter_width, \
-            conv_filters=100, no_ext_feats=False, ext_feats_size=4, n_classes=2, cuda=False):
-        super(QAModel, self).__init__()
-
-        self.no_ext_feats = no_ext_feats
-
-        self.conv_channels = conv_filters
-        n_hidden = 2*self.conv_channels + (0 if no_ext_feats else ext_feats_size)
-
-        self.conv_q = nn.Sequential(
-            nn.Conv1d(input_n_dim, self.conv_channels, filter_width, padding=filter_width-1),
-            nn.Tanh()
-        )
-
-        self.conv_a = nn.Sequential(
-            nn.Conv1d(input_n_dim, self.conv_channels, filter_width, padding=filter_width-1),
-            nn.Tanh()
-        )
-
-        self.combined_feature_vector = nn.Linear(2*self.conv_channels + \
-            (0 if no_ext_feats else ext_feats_size), n_hidden)
-
-        self.combined_features_activation = nn.Tanh()
-        self.dropout = nn.Dropout(0.5)
-        self.hidden = nn.Linear(n_hidden, n_classes)
-        self.logsoftmax = nn.LogSoftmax()
-
-        if cuda and torch.cuda.is_available():
-            self.conv_q, self.conv_a = self.conv_q.cuda(), self.conv_a.cuda()
-            self.combined_feature_vector = self.combined_feature_vector.cuda()
-            self.combined_features_activation = self.combined_features_activation.cuda()
-            self.dropout, self.hidden, self.logsoftmax = self.dropout.cuda(), self.hidden.cuda(), self.logsoftmax.cuda()
-
-    def forward(self, question, answer, ext_feats):
-
-        q = self.conv_q.forward(question)
-        q = F.max_pool1d(q, q.size()[2])
-        q = q.view(-1, self.conv_channels)
-        # logger.debug('forward q: {}'.format(q))
-
-        a = self.conv_a.forward(answer)
-        a = F.max_pool1d(a, a.size()[2])
-        a = a.view(-1, self.conv_channels)
-
-        x = None
-        if self.no_ext_feats:
-            x = torch.cat([q, a], 1)
-            # logger.debug('no_ext_feats')
+        if self.mode == 'multichannel':
+            input_channel = 2
         else:
-            x = torch.cat([q, a, ext_feats], 1)
-            # logger.debug('with ext_feats')
+            input_channel = 1
 
-        # logger.debug('featvec x: {}'.format(x))
-        # logger.debug(x.creator)
+        self.question_embed = nn.Embedding(questions_num, words_dim)
+        self.answer_embed = nn.Embedding(answers_num, words_dim)
+        self.static_question_embed = nn.Embedding(questions_num, words_dim)
+        self.nonstatic_question_embed = nn.Embedding(questions_num, words_dim)
+        self.static_answer_embed = nn.Embedding(answers_num, words_dim)
+        self.nonstatic_answer_embed = nn.Embedding(answers_num, words_dim)
+        self.static_question_embed.weight.requires_grad = False
+        self.static_answer_embed.weight.requires_grad = False
 
-        x = self.combined_feature_vector.forward(x)
-        x = self.combined_features_activation.forward(x)
+        self.conv_q = nn.Conv2d(input_channel, output_channel, (filter_width, words_dim), padding=(filter_width - 1, 0))
+        self.conv_a = nn.Conv2d(input_channel, output_channel, (filter_width, words_dim), padding=(filter_width - 1, 0))
+
+        self.dropout = nn.Dropout(config.dropout)
+        n_hidden = 2 * output_channel + ext_feats_size
+
+        self.combined_feature_vector = nn.Linear(n_hidden, n_hidden)
+        self.hidden = nn.Linear(n_hidden, n_classes)
+
+    def forward(self, x_question, x_answer, x_ext):
+        if self.mode == 'rand':
+            question = self.question_embed(x_question).unsqueeze(1)
+            answer = self.answer_embed(x_answer).unsqueeze(1) # (batch, sent_len, embed_dim)
+            x = [F.tanh(self.conv_q(question)).squeeze(3), F.tanh(self.conv_a(answer)).squeeze(3)]
+            x = [F.max_pool1d(i, i.size(2)).squeeze(2) for i in x]  # max-over-time pooling
+        # actual SM model mode (Severyn & Moschitti, 2015)
+        elif self.mode == 'static':
+            question = self.static_question_embed(x_question).unsqueeze(1)
+            answer = self.static_answer_embed(x_answer).unsqueeze(1) # (batch, sent_len, embed_dim)
+            x = [F.tanh(self.conv_q(question)).squeeze(3), F.tanh(self.conv_a(answer)).squeeze(3)]
+            x = [F.max_pool1d(i, i.size(2)).squeeze(2) for i in x]  # max-over-time pooling
+        elif self.mode == 'non-static':
+            question = self.nonstatic_question_embed(x_question).unsqueeze(1)
+            answer = self.nonstatic_answer_embed(x_answer).unsqueeze(1) # (batch, sent_len, embed_dim)
+            x = [F.tanh(self.conv_q(question)).squeeze(3), F.tanh(self.conv_a(answer)).squeeze(3)]
+            x = [F.max_pool1d(i, i.size(2)).squeeze(2) for i in x]  # max-over-time pooling
+        elif self.mode == 'multichannel':
+            question_static = self.static_question_embed(x_question)
+            answer_static = self.static_answer_embed(x_answer)
+            question_nonstatic = self.nonstatic_question_embed(x_question)
+            answer_nonstatic = self.nonstatic_answer_embed(x_answer)
+            question = torch.stack([question_static, question_nonstatic], dim=1)
+            answer = torch.stack([answer_static, answer_nonstatic], dim=1)
+            x = [F.tanh(self.conv_q(question)).squeeze(3), F.tanh(self.conv_a(answer)).squeeze(3)]
+            x = [F.max_pool1d(i, i.size(2)).squeeze(2) for i in x]  # max-over-time pooling
+        else:
+            print("Unsupported Mode")
+            exit()
+
+        # append external features and feed to fc
+        x.append(x_ext)
+        x = torch.cat(x, 1)
+
+        x = F.tanh(self.combined_feature_vector(x))
         x = self.dropout(x)
         x = self.hidden(x)
-        x = self.logsoftmax(x)
-
         return x
-
-
