+++ sm_cnn/requirements.txt
@@ -1,5 +1,4 @@
-gensim==1.0.1
 nltk==3.2.1
 numpy==1.11.3
-pandas==0.19.2
+gensim==1.0.1
 pytorch==0.1.12
