+++ sm_cnn/train.py
@@ -1,261 +1,205 @@
 import time
-
+import os
 import numpy as np
+import random
+
 import torch
 import torch.nn as nn
-import torch.optim as optim
-from torch.autograd import Variable
+from torchtext import data
 
-import utils
+from args import get_args
+from model import SmPlusPlus
+from trec_dataset import TrecDataset
+from wiki_dataset import WikiDataset
+from evaluate import evaluate
 
-# logging setup
-import logging
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.INFO)
+args = get_args()
+config = args
 
-ch = logging.StreamHandler()
-ch.setLevel(logging.DEBUG)
-formatter = logging.Formatter('%(levelname)s - %(message)s')
-ch.setFormatter(formatter)
-logger.addHandler(ch)
+torch.manual_seed(args.seed)
 
+def set_vectors(field, vector_path):
+    if os.path.isfile(vector_path):
+        stoi, vectors, dim = torch.load(vector_path)
+        field.vocab.vectors = torch.Tensor(len(field.vocab), dim)
 
-class Trainer(object):
+        for i, token in enumerate(field.vocab.itos):
+            wv_index = stoi.get(token, None)
+            if wv_index is not None:
+                field.vocab.vectors[i] = vectors[wv_index]
+            else:
+                # initialize <unk> with U(-0.25, 0.25) vectors
+                field.vocab.vectors[i] = torch.FloatTensor(dim).uniform_(-0.25, 0.25)
+    else:
+        print("Error: Need word embedding pt file")
+        exit(1)
+    return field
 
-    def __init__(self, model, eta, mom, no_loss_reg, vec_dim, cuda=False):
-        # set the random seeds for every instance of trainer.
-        # needed to ensure reproduction of random word vectors for out of vocab terms
-        torch.manual_seed(1234)
-        np.random.seed(1234)
-        self.cuda = cuda
-        self.unk_term = np.random.uniform(-0.25, 0.25, vec_dim)
+# Set default configuration in : args.py
+args = get_args()
+config = args
 
-        self.reg = 1e-5
-        self.no_loss_reg = no_loss_reg
-        self.model = model
-        self.criterion = nn.CrossEntropyLoss()
-        #self.criterion = nn.NLLLoss()
-        self.optimizer = optim.SGD(self.model.parameters(), lr=eta, momentum=mom, \
-            weight_decay=(0 if no_loss_reg else self.reg))
+# Set random seed for reproducibility
+torch.manual_seed(args.seed)
+if not args.cuda:
+    args.gpu = -1
+if torch.cuda.is_available() and args.cuda:
+    print("Note: You are using GPU for training")
+    torch.cuda.set_device(args.gpu)
+    torch.cuda.manual_seed(args.seed)
+if torch.cuda.is_available() and not args.cuda:
+    print("You have Cuda but you're using CPU for training.")
+np.random.seed(args.seed)
+random.seed(args.seed)
 
-        self.data_splits = {}
-        self.embeddings = {}
-        self.vec_dim = vec_dim
+QID = data.Field(sequential=False)
+QUESTION = data.Field(batch_first=True)
+ANSWER = data.Field(batch_first=True)
+LABEL = data.Field(sequential=False)
+EXTERNAL = data.Field(sequential=True, tensor_type=torch.FloatTensor, batch_first=True, use_vocab=False,
+            postprocessing=data.Pipeline(lambda arr, _, train: [float(y) for y in arr]))
 
+if config.dataset == 'TREC':
+    train, dev, test = TrecDataset.splits(QID, QUESTION, ANSWER, EXTERNAL, LABEL)
+elif config.dataset == 'wiki':
+    train, dev, test = WikiDataset.splits(QID, QUESTION, ANSWER, EXTERNAL, LABEL)
+else:
+    print("Unsupported dataset")
+    exit()
 
-    def load_input_data(self, dataset_root_folder, word_vectors_cache_file, \
-            train_set_folder, dev_set_folder, test_set_folder, load_ext_feats=True):
-        for set_folder in [test_set_folder, dev_set_folder, train_set_folder]:
-            if set_folder:
-                questions, sentences, labels, maxlen_q, maxlen_s, vocab = \
-                    utils.read_in_dataset(dataset_root_folder, set_folder)
-
-                self.data_splits[set_folder] = [questions, sentences, labels, maxlen_q, maxlen_s]
-
-                default_ext_feats = [np.zeros(4)] * len(self.data_splits[set_folder][0])
-                self.data_splits[set_folder].append(default_ext_feats)
-
-                utils.load_cached_embeddings(word_vectors_cache_file, vocab, self.embeddings,
-                                             [] if "train" in set_folder else self.unk_term)
-
-
-    def regularize_loss(self, loss):
-
-        flattened_params = []
-
-        for p in self.model.parameters():
-            f = p.data.clone()
-            flattened_params.append(f.view(-1))
-
-        fp = torch.cat(flattened_params)
-
-        loss = loss + 0.5 * self.reg * fp.norm() * fp.norm()
-
-        # for p in self.model.parameters():
-        #     loss = loss + 0.5 * self.reg * p.norm() * p.norm()
-
-        return loss
+QID.build_vocab(train, dev, test)
+QUESTION.build_vocab(train, dev, test)
+ANSWER.build_vocab(train, dev, test)
+LABEL.build_vocab(train, dev, test)
 
 
-    def _train(self, xq, xa, ext_feats, ys):
+QUESTION = set_vectors(QUESTION, args.vector_cache)
+ANSWER = set_vectors(ANSWER, args.vector_cache)
 
-        self.optimizer.zero_grad()
-        output = self.model(xq, xa, ext_feats)
-        loss = self.criterion(output, ys)
-        # logger.debug('loss after criterion {}'.format(loss))
+train_iter = data.Iterator(train, batch_size=args.batch_size, device=args.gpu, train=True, repeat=False,
+                                   sort=False, shuffle=True)
+dev_iter = data.Iterator(dev, batch_size=args.batch_size, device=args.gpu, train=False, repeat=False,
+                                   sort=False, shuffle=False)
+test_iter = data.Iterator(test, batch_size=args.batch_size, device=args.gpu, train=False, repeat=False,
+                                   sort=False, shuffle=False)
 
-        # NOTE: regularizing location 1
-        if not self.no_loss_reg:
-            loss = self.regularize_loss(loss)
-        #     logger.debug('loss after regularizing {}'.format(loss))
+config.target_class = len(LABEL.vocab)
+config.questions_num = len(QUESTION.vocab)
+config.answers_num = len(ANSWER.vocab)
 
+print("Dataset {}    Mode {}".format(args.dataset, args.mode))
+print("VOCAB num", len(QUESTION.vocab))
+print("LABEL.target_class:", len(LABEL.vocab))
+print("LABELS:", LABEL.vocab.itos)
+print("Train instance", len(train))
+print("Dev instance", len(dev))
+print("Test instance", len(test))
+
+if args.resume_snapshot:
+    if args.cuda:
+        model = torch.load(args.resume_snapshot, map_location=lambda storage, location: storage.cuda(args.gpu))
+    else:
+        model = torch.load(args.resume_snapshot, map_location=lambda storage, location: storage)
+else:
+    model = SmPlusPlus(config)
+    model.static_question_embed.weight.data.copy_(QUESTION.vocab.vectors)
+    model.nonstatic_question_embed.weight.data.copy_(QUESTION.vocab.vectors)
+    model.static_answer_embed.weight.data.copy_(ANSWER.vocab.vectors)
+    model.nonstatic_answer_embed.weight.data.copy_(ANSWER.vocab.vectors)
+
+    if args.cuda:
+        model.cuda()
+        print("Shift model to GPU")
+
+
+parameter = filter(lambda p: p.requires_grad, model.parameters())
+
+# the SM model originally follows SGD but Adadelta is used here
+optimizer = torch.optim.Adadelta(parameter, lr=args.lr, weight_decay=args.weight_decay)
+criterion = nn.CrossEntropyLoss()
+early_stop = False
+best_dev_map = 0
+iterations = 0
+iters_not_improved = 0
+epoch = 0
+start = time.time()
+header = '  Time Epoch Iteration Progress    (%Epoch)   Loss   Dev/Loss     Accuracy  Dev/Accuracy'
+dev_log_template = ' '.join('{:>6.0f},{:>5.0f},{:>9.0f},{:>5.0f}/{:<5.0f} {:>7.0f}%,{:>8.6f},{:8.6f},{:12.4f},{:12.4f}'.split(','))
+log_template = ' '.join('{:>6.0f},{:>5.0f},{:>9.0f},{:>5.0f}/{:<5.0f} {:>7.0f}%,{:>8.6f},{},{:12.4f},{}'.split(','))
+os.makedirs(args.save_path, exist_ok=True)
+os.makedirs(os.path.join(args.save_path, args.dataset), exist_ok=True)
+print(header)
+
+index2label = np.array(LABEL.vocab.itos)
+index2qid = np.array(QID.vocab.itos)
+index2question = np.array(ANSWER.vocab.itos)
+
+while True:
+    if early_stop:
+        print("Early Stopping. Epoch: {}, Best Dev Acc: {}".format(epoch, best_dev_map))
+        break
+    epoch += 1
+    train_iter.init_epoch()
+    n_correct, n_total = 0, 0
+
+    for batch_idx, batch in enumerate(train_iter):
+        iterations += 1
+        model.train(); optimizer.zero_grad()
+        scores = model(batch.question, batch.answer, batch.ext_feat)
+        n_correct += (torch.max(scores, 1)[1].view(batch.label.size()).data == batch.label.data).sum()
+        n_total += batch.batch_size
+        train_acc = 100. * n_correct / n_total
+
+        loss = criterion(scores, batch.label)
         loss.backward()
+        optimizer.step()
 
-        # logger.debug('AFTER backward')
-        #logger.debug('params {}'.format([p for p in self.model.parameters()]))
-        # logger.debug('params grads {}'.format([p.grad for p in self.model.parameters()]))
+        # Evaluate performance on validation set
+        if iterations % args.dev_every == 1:
+            # switch model into evaluation mode
+            model.eval()
+            dev_iter.init_epoch()
+            n_dev_correct = 0
+            dev_losses = []
+            instance = []
+            for dev_batch_idx, dev_batch in enumerate(dev_iter):
+                qid_array = index2qid[np.transpose(dev_batch.qid.cpu().data.numpy())]
+                true_label_array = index2label[np.transpose(dev_batch.label.cpu().data.numpy())]
 
-        # NOTE: regularizing location 2. It would seem that location 1 is correct?
-        #if not self.no_loss_reg:
-        #    loss = self.regularize_loss(loss)
-            # logger.debug('loss after regularizing {}'.format(loss))
-
-        self.optimizer.step()
-
-        # logger.debug('AFTER step')
-        #logger.debug('params {}'.format([p for p in self.model.parameters()]))
-        # logger.debug('params grads {}'.format([p.grad for p in self.model.parameters()]))
-
-        return loss.data[0], self.pred_equals_y(output, ys)
-
-
-    def pred_equals_y(self, pred, y):
-        _, best = pred.max(1)
-        best = best.data.long().squeeze()
-        return torch.sum(y.data.long() == best)
-
-
-    def test(self, set_folder, batch_size):
-        logger.info('----- Predictions on {} '.format(set_folder))
-
-        questions, sentences, labels, maxlen_q, maxlen_s, ext_feats = \
-            self.data_splits[set_folder]
-        word_vectors, vec_dim = self.embeddings, self.vec_dim
-
-        self.model.eval()
-
-        batch_size = 1
-
-        total_loss = 0.0
-        total_correct = 0.0
-        num_batches = np.ceil(len(questions)/batch_size)
-        y_pred = np.zeros(len(questions))
-        ypc = 0
-
-        for k in range(int(num_batches)):
-            batch_start = k * batch_size
-            batch_end = (k+1) * batch_size
-            # convert raw questions and sentences to tensors
-            batch_inputs, batch_labels = self.get_tensorized_inputs(
-                questions[batch_start:batch_end],
-                sentences[batch_start:batch_end],
-                labels[batch_start:batch_end],
-                ext_feats[batch_start:batch_end],
-                word_vectors, vec_dim
-            )
-
-            xq, xa, x_ext_feats = batch_inputs[0]
-            y = batch_labels[0]
-
-            pred = self.model(xq, xa, x_ext_feats)
-            loss = self.criterion(pred, y)
-            pred = torch.exp(pred)
-            total_loss += loss
-            # total_correct += self.pred_equals_y(pred, y)
-
-            y_pred[ypc] = pred.data.squeeze()[1]
-            # ^ we want to score for relevance, NOT the predicted class
-            ypc += 1
-
-        # logger.info('{}_correct {}'.format(set_folder, total_correct))
-        # logger.info('{}_loss {}'.format(set_folder, total_loss.data[0]))
-        logger.info('{} total {}'.format(set_folder, len(labels)))
-        # logger.info('{}_loss = {:.4f}, acc = {:.4f}'.format(set_folder, total_loss.data[0]/len(labels), float(total_correct)/len(labels))
-        #logger.info('{}_loss = {:.4f}'.format(set_folder, total_loss.data[0]/len(labels)))
-
-        return y_pred
+                scores = model(dev_batch.question, dev_batch.answer, dev_batch.ext_feat)
+                n_dev_correct += (torch.max(scores, 1)[1].view(dev_batch.label.size()).data == dev_batch.label.data).sum()
+                dev_loss = criterion(scores, dev_batch.label)
+                dev_losses.append(dev_loss.data[0])
+                index_label = np.transpose(torch.max(scores, 1)[1].view(dev_batch.label.size()).cpu().data.numpy())
+                label_array = index2label[index_label]
+                # get the relevance scores
+                score_array = scores[:, 2].cpu().data.numpy()
+                for i in range(dev_batch.batch_size):
+                    this_qid, predicted_label, score, gold_label = qid_array[i], label_array[i], score_array[i], true_label_array[i]
+                    instance.append((this_qid, predicted_label, score, gold_label))
 
 
-    def train(self, set_folder, batch_size, debug_single_batch):
-        train_start_time = time.time()
+            dev_map, dev_mrr = evaluate(instance, config.dataset, 'valid', config.mode)
+            print(dev_log_template.format(time.time() - start,
+                                          epoch, iterations, 1 + batch_idx, len(train_iter),
+                                          100. * (1 + batch_idx) / len(train_iter), loss.data[0],
+                                          sum(dev_losses) / len(dev_losses), train_acc, dev_map))
 
-        questions, sentences, labels, maxlen_q, maxlen_s, ext_feats = \
-            self.data_splits[set_folder]
-        word_vectors, vec_dim = self.embeddings, self.vec_dim
+            # Update validation results
+            if dev_map > best_dev_map:
+                iters_not_improved = 0
+                best_dev_map = dev_map
+                snapshot_path = os.path.join(args.save_path, args.dataset, args.mode+'_best_model.pt')
+                torch.save(model, snapshot_path)
+            else:
+                iters_not_improved += 1
+                if iters_not_improved >= args.patience:
+                    early_stop = True
+                    break
 
-        # set model for training modep
-        self.model.train()
-
-        train_loss, train_correct = 0., 0.
-        num_batches = np.ceil(len(questions)/float(batch_size))
-
-        for k in range(int(num_batches)):
-            batch_start = k * batch_size
-            batch_end = (k+1) * batch_size
-
-            # convert raw questions and sentences to tensors
-            batch_inputs, batch_labels = self.get_tensorized_inputs(
-                questions[batch_start:batch_end],
-                sentences[batch_start:batch_end],
-                labels[batch_start:batch_end],
-                ext_feats[batch_start:batch_end],
-                word_vectors, vec_dim
-            )
-
-            xq, xa, x_ext_feats = batch_inputs[0]
-
-            ys = batch_labels[0]
-
-            batch_loss, batch_correct = self._train(xq, xa, x_ext_feats, ys)
-
-            # logger.debug('batch_loss {}, batch_correct {}'.format(batch_loss, batch_correct))
-            train_loss += batch_loss
-            # train_correct += batch_correct
-            if debug_single_batch:
-                break
-
-        # logger.info('train_correct {}'.format(train_correct))
-        logger.info('train_loss {}'.format(train_loss))
-        logger.info('total training batches = {}'.format(num_batches))
-        logger.info('train_loss = {:.4f}'.format(
-            train_loss/num_batches
-        ))
-        logger.info('training time = {:.3f} seconds'.format(time.time() - train_start_time))
-        return train_correct/num_batches
-
-
-    def make_input_matrix(self, sentence, word_vectors, vec_dim):
-        terms = sentence.strip().split()[:60]
-        # NOTE: we are truncating the inputs to 60 words.
-
-        word_embeddings = torch.zeros(len(terms), vec_dim).type(torch.DoubleTensor)
-        for i in range(len(terms)):
-            word = terms[i]
-            emb = torch.from_numpy(word_vectors[word])
-            word_embeddings[i] = emb
-
-        input_tensor = torch.zeros(1, vec_dim, len(terms))
-        input_tensor[0] = torch.transpose(word_embeddings, 0, 1)
-        if self.cuda and torch.cuda.is_available():
-            input_tensor = input_tensor.cuda()
-        return input_tensor
-
-
-    def get_tensorized_inputs(self, batch_ques, batch_sents, batch_labels, batch_ext_feats, \
-            word_vectors, vec_dim):
-        batch_size = len(batch_ques)
-        # NOTE: ideal batch size is one, because sentences are all of different length.
-        # In other words, we have no option but to feed in sentences one by one into the model
-        # and compute loss at the end.
-
-        # TODO: what if the sentences in a batch are all of different lengths?
-        # - should be have the longest sentence as 2nd dim?
-        #   - would zero endings work for other smaller sentences?
-
-        y = torch.LongTensor(batch_size).type(torch.LongTensor)
-        if self.cuda and torch.cuda.is_available():
-            y = y.cuda()
-
-        tensorized_inputs = []
-        for i in range(len(batch_ques)):
-            xq = Variable(self.make_input_matrix(batch_ques[i], word_vectors, vec_dim))
-            xs = Variable(self.make_input_matrix(batch_sents[i], word_vectors, vec_dim))
-            ext_feats = torch.FloatTensor(batch_ext_feats[i])
-            if self.cuda and torch.cuda.is_available():
-                ext_feats = ext_feats.cuda()
-            ext_feats = Variable(ext_feats)
-            ext_feats = torch.unsqueeze(ext_feats, 0)
-            y[i] = batch_labels[i]
-            tensorized_inputs.append((xq, xs, ext_feats))
-
-        return tensorized_inputs, Variable(y)
+        if iterations % args.log_every == 1:
+            # print progress message
+            print(log_template.format(time.time() - start,
+                                      epoch, iterations, 1 + batch_idx, len(train_iter),
+                                      100. * (1 + batch_idx) / len(train_iter), loss.data[0], ' ' * 8,
+                                      n_correct / n_total * 100, ' ' * 12))
