+++ sm_cnn/trec_dataset.py
@@ -5,12 +5,11 @@ class TrecDataset(data.TabularDataset):
     dirname = 'data'
     @classmethod
 
-    def splits(cls, question_id, question_field, answer_field, external_field, label_field,
-               train='train.tsv', validation='dev.tsv', test='test.tsv'):
+    def splits(cls, question_id, question_field, answer_field, external_field, label_field, root='.data',
+               train='trecqa.train.tsv', validation='trecqa.dev.tsv', test='trecqa.test.tsv'):
         path = './data'
-        prefix_name = 'trecqa.'
         return super(TrecDataset, cls).splits(
-            os.path.join(path, prefix_name), train, validation, test,
+            path, root, train, validation, test,
             format='TSV', fields=[('qid', question_id), ('label', label_field), ('question', question_field),
                                   ('answer', answer_field), ('ext_feat', external_field)]
         )
