--- sm_modified_cnn/.gitignore
@@ -1,6 +1,0 @@
-*pyc
-*.pt
-text/
-trained_models/
-trec_eval-8.0/trec_eval.dSYM
-data/
