--- sm_modified_cnn/requirements.txt
@@ -1,4 +1,0 @@
-nltk==3.2.1
-numpy==1.11.3
-gensim==1.0.1
-pytorch==0.1.12
